# swagger.json
{"openapi":"3.1.0","info":{"title":"Rural API","description":"Sistema de cadastro e monitoramento de produtores rurais","version":"1.0.0"},"paths":{"/auth/register":{"post":{"tags":["Auth"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/token":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/producers/":{"get":{"tags":["Producers"],"summary":"Get Producers","operationId":"get_producers_producers__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProducerOut"},"type":"array","title":"Response Get Producers Producers  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Producers"],"summary":"Add Producer","operationId":"add_producer_producers__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProducerCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProducerOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/producers/{producer_id}":{"put":{"tags":["Producers"],"summary":"Edit Producer","operationId":"edit_producer_producers__producer_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"producer_id","in":"path","required":true,"schema":{"type":"integer","title":"Producer Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProducerCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProducerOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Producers"],"summary":"Remove Producer","operationId":"remove_producer_producers__producer_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"producer_id","in":"path","required":true,"schema":{"type":"integer","title":"Producer Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/properties/{producer_id}":{"post":{"tags":["Properties"],"summary":"Add Property","operationId":"add_property_properties__producer_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"producer_id","in":"path","required":true,"schema":{"type":"integer","title":"Producer Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PropertyCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PropertyOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/harvests/{property_id}":{"post":{"tags":["Harvests"],"summary":"Add Harvest","operationId":"add_harvest_harvests__property_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"property_id","in":"path","required":true,"schema":{"type":"integer","title":"Property Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HarvestCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HarvestOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/crops/{harvest_id}":{"post":{"tags":["Crops"],"summary":"Add Crop","operationId":"add_crop_crops__harvest_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"harvest_id","in":"path","required":true,"schema":{"type":"integer","title":"Harvest Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CropCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CropOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/dashboard/":{"get":{"tags":["Dashboard"],"summary":"Get Dashboard","operationId":"get_dashboard_dashboard__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Body_login_auth_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_token_post"},"CropCreate":{"properties":{"culture_name":{"type":"string","title":"Culture Name"}},"type":"object","required":["culture_name"],"title":"CropCreate"},"CropOut":{"properties":{"culture_name":{"type":"string","title":"Culture Name"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["culture_name","id"],"title":"CropOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HarvestCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"HarvestCreate"},"HarvestOut":{"properties":{"name":{"type":"string","title":"Name"},"id":{"type":"integer","title":"Id"},"crops":{"items":{"$ref":"#/components/schemas/CropOut"},"type":"array","title":"Crops","default":[]}},"type":"object","required":["name","id"],"title":"HarvestOut"},"ProducerCreate":{"properties":{"name":{"type":"string","title":"Name"},"cpf_cnpj":{"type":"string","title":"Cpf Cnpj"}},"type":"object","required":["name","cpf_cnpj"],"title":"ProducerCreate"},"ProducerOut":{"properties":{"name":{"type":"string","title":"Name"},"cpf_cnpj":{"type":"string","title":"Cpf Cnpj"},"id":{"type":"integer","title":"Id"},"properties":{"items":{"$ref":"#/components/schemas/PropertyOut"},"type":"array","title":"Properties","default":[]}},"type":"object","required":["name","cpf_cnpj","id"],"title":"ProducerOut"},"PropertyCreate":{"properties":{"name":{"type":"string","title":"Name"},"city":{"type":"string","title":"City"},"state":{"type":"string","title":"State"},"total_area":{"type":"number","title":"Total Area"},"agricultural_area":{"type":"number","title":"Agricultural Area"},"vegetation_area":{"type":"number","title":"Vegetation Area"}},"type":"object","required":["name","city","state","total_area","agricultural_area","vegetation_area"],"title":"PropertyCreate"},"PropertyOut":{"properties":{"name":{"type":"string","title":"Name"},"city":{"type":"string","title":"City"},"state":{"type":"string","title":"State"},"total_area":{"type":"number","title":"Total Area"},"agricultural_area":{"type":"number","title":"Agricultural Area"},"vegetation_area":{"type":"number","title":"Vegetation Area"},"id":{"type":"integer","title":"Id"},"harvests":{"items":{"$ref":"#/components/schemas/HarvestOut"},"type":"array","title":"Harvests","default":[]}},"type":"object","required":["name","city","state","total_area","agricultural_area","vegetation_area","id"],"title":"PropertyOut"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserCreate"},"UserOut":{"properties":{"username":{"type":"string","title":"Username"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["username","id"],"title":"UserOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/token"}}}}}}